requirements.txt
mediapipe
opencv-python
pillow
numpy
pandas
scikit-learn
matplotlib
torch
torchvision
